moved methods out of HomeController to their apprp controllers
