+++ src/main/java/edu/msudenver/cs3250/group6/msubanner/controllers/HomeController.java
@@ -1,16 +1,9 @@
 package edu.msudenver.cs3250.group6.msubanner.controllers;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
-
 import edu.msudenver.cs3250.group6.msubanner.Global;
-import edu.msudenver.cs3250.group6.msubanner.entities.Days;
-import edu.msudenver.cs3250.group6.msubanner.services.BuildingService;
-import edu.msudenver.cs3250.group6.msubanner.services.RoomService;
-import edu.msudenver.cs3250.group6.msubanner.services.SemesterService;
 
 /*
  *  The purpose here is to get a mapping to the index.html in templates
@@ -26,18 +19,6 @@ import edu.msudenver.cs3250.group6.msubanner.services.SemesterService;
 @Controller
 class HomeController {
 
-    /** The building repository. */
-    @Autowired
-    private BuildingService buildingService;
-
-    /** The room service. */
-    @Autowired
-    private RoomService roomService;
-
-    /** The semester service. */
-    @Autowired
-    private SemesterService semesterService;
-
     /**
      * Maps the home page to index.html.
      *
@@ -49,34 +30,4 @@ class HomeController {
         mav.addObject("school_name", Global.SCHOOL_NAME);
         return mav;
     }
-
-    /**
-     * Maps to the add department form.
-     *
-     * @return the add department form string
-     */
-    @RequestMapping("/departments/adddepartment")
-    ModelAndView addDepartmentForm() {
-        ModelAndView mav = new ModelAndView("adddepartmentform");
-        mav.addObject("school_name", Global.SCHOOL_NAME);
-        return mav;
-    }
-
-    /**
-     * Maps to the add section form.
-     *
-     * @return the add section form string
-     */
-    @RequestMapping("schedules/addschedule")
-    ModelAndView addScheduleForm() {
-        ModelAndView mav = new ModelAndView("addscheduleform");
-        mav.addObject("allrooms", roomService.getAllRooms());
-        mav.addObject("allbuildings", buildingService.getAllBuildings());
-        mav.addObject("allsemesters", semesterService.getAllSemesters());
-        mav.addObject("school_name", Global.SCHOOL_NAME);
-        mav.addObject("alldays", Days.Day.values());
-        return mav;
-    }
-
-
 }
