+++ src/main/java/edu/msudenver/cs3250/group6/msubanner/controllers/SectionController.java
@@ -202,4 +202,5 @@ public class SectionController {
         return mav;
     }
 
+
 }
