--- src/test/java/edu/msudenver/cs3250/group6/msubanner/HomeControllerTest.java
@@ -1,5 +1,0 @@
-package edu.msudenver.cs3250.group6.msubanner;
-
-public class HomeControllerTest {
-
-}
