+++ src/test/java/edu/msudenver/cs3250/group6/msubanner/controllers/HomeControllerTest.java
@@ -1,0 +1,5 @@
+package edu.msudenver.cs3250.group6.msubanner.controllers;
+
+public class HomeControllerTest {
+
+}
